requirements.txt
arabic-reshaper==2.1.4
asgiref==3.6.0
asn1crypto==1.5.1
beautifulsoup4==4.11.1
Brotli==1.0.9
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
crispy-bootstrap5==0.7
cryptography==38.0.4
cssselect2==0.7.0
Django==4.1.4
django-bootstrap4==22.3
django-bootstrap5==22.2
django-crispy-forms==1.14.0
django-filter==22.1
django-flatpickr==2.0.0
django-tables2==2.4.1
django-weasyprint==2.2.0
fonttools==4.38.0
future==0.18.2
html5lib==1.1
idna==3.4
lxml==4.9.2
mysqlclient==2.1.1
oscrypto==1.3.0
phonenumberslite==8.13.2
Pillow==9.3.0
pycparser==2.21
pydantic==1.10.2
pydyf==0.5.0
pyHanko==0.16.0
pyhanko-certvalidator==0.19.8
PyPDF3==1.0.6
pyphen==0.13.2
python-bidi==0.4.2
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
reportlab==3.6.12
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
svglib==1.4.1
tinycss2==1.2.1
tqdm==4.64.1
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
uritools==4.0.0
urllib3==1.26.13
weasyprint==57.2
webencodings==0.5.1
zopfli==0.2.2
